skip making plots and drop samples

+++ R/epinow.R
@@ -60,7 +60,8 @@ epinow <- function(reported_cases, model, samples = 1000, stan_args,
                    burn_in = 0, stationary = FALSE, fixed = FALSE, fixed_future_rt = FALSE,
                    future = FALSE, max_execution_time = Inf, prior_smoothing_window = 7,
                    return_fit = FALSE, forecast_model, ensemble_type = "mean",
-                   return_estimates = TRUE, target_folder, target_date, verbose = FALSE) {
+                   return_estimates = TRUE, keep_samples = TRUE, make_plots = TRUE, 
+                   target_folder, target_date, verbose = FALSE) {
 
   if (!return_estimates & missing(target_folder)) {
     futile.logger::flog.fatal("Either return estimates or save to a target folder")
@@ -132,7 +133,9 @@ epinow <- function(reported_cases, model, samples = 1000, stan_args,
 
   # Report estimates --------------------------------------------------------
   if (!is.null(target_folder)) {
-    saveRDS(estimates$samples, paste0(target_folder, "/estimate_samples.rds"))
+    if (keep_samples) {
+      saveRDS(estimates$samples, paste0(target_folder, "/estimate_samples.rds"))
+    }
     saveRDS(estimates$summarised, paste0(target_folder, "/summarised_estimates.rds"))
 
     if (return_fit) {
@@ -153,15 +156,19 @@ epinow <- function(reported_cases, model, samples = 1000, stan_args,
   }
   # Report cases ------------------------------------------------------------
   if (!missing(forecast_model) & !is.null(target_folder)) {
-    saveRDS(forecast$samples, paste0(target_folder, "/forecast_samples.rds"))
+    if (keep_samples){
+      saveRDS(forecast$samples, paste0(target_folder, "/forecast_samples.rds"))
+    }
     saveRDS(forecast$summarised, paste0(target_folder, "/summarised_forecast.rds"))
   }
   # Report forcasts ---------------------------------------------------------
 
   if (missing(forecast_model)) {
     estimated_reported_cases <- list()
-    estimated_reported_cases$samples <- estimates$samples[variable == "reported_cases"][,
-      .(date, sample, cases = value, type = "gp_rt")]
+    if (keep_samples) {
+      estimated_reported_cases$samples <- estimates$samples[variable == "reported_cases"][,
+                                          .(date, sample, cases = value, type = "gp_rt")]
+    }
     estimated_reported_cases$summarised <- estimates$summarised[variable == "reported_cases"][,
       type := "gp_rt"][, variable := NULL][, strat := NULL]
   }else {
@@ -176,19 +183,22 @@ epinow <- function(reported_cases, model, samples = 1000, stan_args,
       return(reported_cases)
     }
 
-    reported_cases_rt <- report_cases_with_forecast(model = "rt")
-    reported_cases_cases <- report_cases_with_forecast(model = "case")
-    reported_cases_ensemble <- report_cases_with_forecast(model = "ensemble")
-
     estimated_reported_cases <- list()
-
-    estimated_reported_cases$samples <- data.table::rbindlist(list(
-      reported_cases_rt$samples[, type := "rt"],
-      reported_cases_cases$samples[, type := "case"],
-      reported_cases_ensemble$samples[, type := "ensemble"],
-      estimates$samples[variable == "reported_cases"][,
-        .(date, sample, cases = value, type = "gp_rt")]
-    ), use.names = TRUE)
+    
+    if (keep_samples) {
+      reported_cases_rt <- report_cases_with_forecast(model = "rt")
+      reported_cases_cases <- report_cases_with_forecast(model = "case")
+      reported_cases_ensemble <- report_cases_with_forecast(model = "ensemble")
+      
+      estimated_reported_cases$samples <- data.table::rbindlist(list(
+        reported_cases_rt$samples[, type := "rt"],
+        reported_cases_cases$samples[, type := "case"],
+        reported_cases_ensemble$samples[, type := "ensemble"],
+        estimates$samples[variable == "reported_cases"][,
+                                                        .(date, sample, cases = value, type = "gp_rt")]
+      ), use.names = TRUE)
+      
+    }
 
     estimated_reported_cases$summarised <- data.table::rbindlist(list(
       reported_cases_rt$summarised[, type := "rt"],
@@ -200,7 +210,9 @@ epinow <- function(reported_cases, model, samples = 1000, stan_args,
   }
 
   if (!is.null(target_folder)) {
-    saveRDS(estimated_reported_cases$samples, paste0(target_folder, "/estimated_reported_cases_samples.rds"))
+    if (keep_samples) {
+      saveRDS(estimated_reported_cases$samples, paste0(target_folder, "/estimated_reported_cases_samples.rds"))
+    }
     saveRDS(estimated_reported_cases$summarised, paste0(target_folder, "/summarised_estimated_reported_cases.rds"))
   }
 
@@ -216,9 +228,10 @@ epinow <- function(reported_cases, model, samples = 1000, stan_args,
   }
 
   #  # Plot --------------------------------------------------------------------
-
-  plots <- report_plots(summarised_estimates = estimates$summarised,
-                        reported = reported_cases, target_folder = target_folder)
+  if (make_plots) {
+    plots <- report_plots(summarised_estimates = estimates$summarised,
+                          reported = reported_cases, target_folder = target_folder)
+  }
 
   # Copy all results to latest folder ---------------------------------------
   if (!is.null(target_folder)) {
@@ -248,7 +261,9 @@ epinow <- function(reported_cases, model, samples = 1000, stan_args,
 
     out$estimated_reported_cases <- estimated_reported_cases
     out$summary <- summary
-    out$plots <- plots
+    if (make_plots) {
+      out$plots <- plots
+    }
     return(out)
   }else {
     return(invisible(NULL))
